Resolve ambiguoud sqrt call (VS8.0)
git-svn-id: https://ferma.svn.sf.net/svnroot/ferma/trunk@840 d51c77f7-620d-0410-a109-9a2f07eaa567

+++ FermaNext/tests/DijkstraTest/DijkstraTest.cpp
@@ -26,7 +26,7 @@ struct Vector
 
 float vectorlength ( const Vector& v )
 {
-    return sqrt(v.x * v.x + v.y * v.y);
+    return sqrt( float(v.x * v.x + v.y * v.y) );
 }
 
 float angle ( const Vector& a, const Vector& b )
